requirements.txt

requests==2.31.0
future>=0.16.0
certifi
tornado>=5.1
python-dotenv